requirements.txt
grpcio==1.18.0
grpcio-tools==1.18.0
kazoo==2.6.0